use-cases/pydantic-ai/examples/gmail_summarizer_agent/requirements.txt
pydantic-ai==0.7.4
python-dotenv>=1.0,<2.0
google-api-python-client>=2.137.0,<3.0
google-auth>=2.29.0,<3.0
google-auth-oauthlib>=1.2.0,<2.0
